Refactor code to achieve 'Presentable state' milestone.
