+++ correctme/bk_tree.py
@@ -1,8 +1,6 @@
 """ BK Tree implementation """
 
 from itertools import imap, ifilter
-import levenshtein_distance as ld
-import time
 
 
 class BKTree:
@@ -50,14 +48,15 @@ class BKTree:
         ascending order of distance.
 
         """
+
         def rec(parent):
             pword, children = parent
             d = self.distfn(word, pword)
             results = []
             if d <= n:
-                results.append( (d, pword) )
+                results.append((d, pword))
 
-            for i in range(d-n, d+n+1):
+            for i in range(d - n, d + n + 1):
                 child = children.get(i)
                 if child is not None:
                     results.extend(rec(child))
@@ -65,7 +64,6 @@ class BKTree:
 
         # sort by distance
         return sorted(rec(self.tree))
-
 
 
 def brute_query(word, words, distfn, n):
@@ -83,44 +81,31 @@ def brute_query(word, words, distfn, n):
 
 
 def maxdepth(tree, count=0):
-    _, children = t
+    _, children = tree
     if len(children):
-        return max(maxdepth(i, c+1) for i in children.values())
+        return max(maxdepth(i, count + 1) for i in children.values())
     else:
-        return c
+        return count
+
 
 # http://en.wikibooks.org/wiki/Algorithm_implementation/Strings/Levenshtein_distance#Python
 
 def levenshtein(s, t):
     m, n = len(s), len(t)
-    d = [range(n+1)]
-    d += [[i] for i in range(1,m+1)]
-    for i in range(0,m):
-        for j in range(0,n):
+    d = [range(n + 1)]
+    d += [[i] for i in range(1, m + 1)]
+    for i in range(0, m):
+        for j in range(0, n):
             cost = 1
             if s[i] == t[j]: cost = 0
 
-            d[i+1].append( min(d[i][j+1]+1, # deletion
-                               d[i+1][j]+1, #insertion
-                               d[i][j]+cost) #substitution
-                           )
+            d[i + 1].append(min(d[i][j + 1] + 1,  # deletion
+                                d[i + 1][j] + 1,  # insertion
+                                d[i][j] + cost)  # substitution
+                            )
     return d[m][n]
 
 
-def dict_words(dictfile="./google-10000-english-no-swears.txt"):
+def dict_words(dictfile):
     "Return an iterator that produces words in the given dictionary."
     return ifilter(len, imap(str.strip, open(dictfile)))
-
-
-# if __name__ == "__main__":
-
-#     t = time.time()
-#     tree = BKTree(ld.get_edit_distance, dict_words('./google-10000-english-no-swears.txt'))
-#     print tree.query("more", 1)
-#     print "time: ", (time.time() - t)
-
-
-#     t = time.time()
-#     tree = BKTree(levenshtein, dict_words('./google-10000-english-no-swears.txt'))
-#     print tree.query("more", 1)
-#     print "time: ", (time.time() - t)
