+++ correctme/edit_distance.py
@@ -1,12 +1,12 @@
-INSERT_COST =  1
-SUBSTITUTION_COST =  2
-DELETE_COST =  1
+INSERT_COST = 1
+SUBSTITUTION_COST = 2
+DELETE_COST = 1
 
 
 def levenshtein_distance(dictionaryString, inputString, prevResult, inputStringPos):
     if inputStringPos >= len(inputString):
         return prevResult[len(dictionaryString)]
-    
+
     newResult = []
     newResult.append(inputStringPos + 1)
 
@@ -14,8 +14,9 @@ def levenshtein_distance(dictionaryString, inputString, prevResult, inputStringP
         if dictionaryString[i - 1] == inputString[inputStringPos]:
             newResult.append(prevResult[i - 1])
         else:
-            newResult.append(min(newResult[i - 1] + INSERT_COST, prevResult[i] + DELETE_COST, prevResult[i - 1] + SUBSTITUTION_COST))
-        
+            newResult.append(
+                min(newResult[i - 1] + INSERT_COST, prevResult[i] + DELETE_COST, prevResult[i - 1] + SUBSTITUTION_COST))
+
     inputStringPos += 1
     return levenshtein_distance(dictionaryString, inputString, newResult, inputStringPos)
 
@@ -25,7 +26,7 @@ def get_edit_distance(dictionaryString, inputString):
     dictionaryString = dictionaryString.lower()
     inputString = inputString.lower()
 
-    if(dictionaryString == inputString):
+    if (dictionaryString == inputString):
         edit_distance = 0
     else:
         prevResult = []
@@ -34,7 +35,3 @@ def get_edit_distance(dictionaryString, inputString):
         edit_distance = levenshtein_distance(dictionaryString, inputString, prevResult, 0)
 
     return edit_distance
-
-
-# if __name__ == "__main__":
-#     print get_edit_distance("gumbo", "gambol")
